requirements.txt
absl-py==2.1.0
ace_tools==0.0
addict==2.4.0
asttokens==2.4.1
attrs==24.1.0
av==12.3.0
backcall==0.2.0
branca==0.8.1
Brotli==1.0.9
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
controlnet-aux==0.0.9
cv==1.0.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.6.7
decorator==5.1.1
diffusers==0.29.2
einops==0.8.0
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.1.0
face-alignment==1.4.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.25
folium==0.20.0
fonttools==4.53.1
fsspec==2024.6.1
gmpy2==2.1.2
grad-cam==1.5.3
grpcio==1.65.4
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.4.0
mmagic==1.0.1
mmcls==1.0.0rc6
mmcv==2.0.1
mmdet==3.0.0
mmengine==0.8.5
mmpretrain==1.0.0rc7
mmsegmentation==1.0.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.7.24
requests==2.32.3
resize-right==0.0.2
rich==13.7.1
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
setuptools==72.2.0
shapely==2.0.5
shiboken6==6.9.2
six==1.16.0
slidingwindow==0.0.14
sounddevice==0.4.7
stack-data==0.6.2
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==0.6.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
triton==2.0.0
ttach==0.0.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.44.0
xyzservices==2025.4.0
yapf==0.40.2
zipp==3.19.2
